+++ apps/web/src/components/superteam/index.ts
@@ -1,4 +1,5 @@
 export * from './v3-components'
+export * from './layout'
 export * from "./team-icon-tile";
 export * from "./team-role";
 export * from "./user-identity";